+++ CrowdControl/src/main/java/com/elmakers/mine/bukkit/plugins/crowd/CrowdControlPlugin.java
@@ -1,10 +1,13 @@
 package com.elmakers.mine.bukkit.plugins.crowd;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.logging.Logger;
 
 import org.bukkit.World;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.CreatureType;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event.Priority;
 import org.bukkit.event.Event.Type;
@@ -13,7 +16,8 @@ import org.bukkit.plugin.PluginDescriptionFile;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 
-import com.elmakers.mine.bukkit.plugins.gameplay.EntityType;
+import com.elmakers.mine.bukkit.plugins.crowd.dao.ControlRule;
+import com.elmakers.mine.bukkit.plugins.crowd.dao.ControlledWorld;
 import com.elmakers.mine.bukkit.plugins.persistence.Persistence;
 import com.elmakers.mine.bukkit.plugins.persistence.PersistencePlugin;
 import com.elmakers.mine.bukkit.plugins.persistence.PluginUtilities;
@@ -71,32 +75,182 @@ public class CrowdControlPlugin extends JavaPlugin
 	    	return;
 	    }
 	    
+	    CrowdControlDefaults d = new CrowdControlDefaults();
 	    utilities = persistence.getUtilities(this);
+	    listener.initialize(persistence, controller);
 	    
-	    crowdCommand = utilities.getPlayerCommand("crowd", "Manage mob spawning", null, PermissionType.ADMINS_ONLY);
-	    crowdControlCommand = crowdCommand.getSubCommand("control", "Control a mob type", "[mob] <chance> <world>", PermissionType.ADMINS_ONLY);
-	    nukeCommand = crowdCommand.getSubCommand("nuke", "Kill all ghasts (or whatever)", "[all | mobtype] [world]", PermissionType.ADMINS_ONLY);
+	    crowdCommand = utilities.getPlayerCommand(d.crowdCommand[0], d.crowdCommand[1], d.crowdCommand[2], PermissionType.ADMINS_ONLY);
+	    crowdControlCommand = crowdCommand.getSubCommand(d.crowdControlCommand[0], d.crowdControlCommand[1], d.crowdControlCommand[2], PermissionType.ADMINS_ONLY);
+	    crowdReleaseCommand = crowdCommand.getSubCommand(d.crowdReleaseCommand[0], d.crowdReleaseCommand[1], d.crowdReleaseCommand[2], PermissionType.ADMINS_ONLY);
+	    nukeCommand = crowdCommand.getSubCommand(d.crowdCommand[0], d.crowdCommand[1], d.crowdCommand[2], PermissionType.ADMINS_ONLY);
 		
 	    crowdControlCommand.bind("onControlCrowd");
+	    crowdReleaseCommand.bind("onReleaseCrowd");
 	    nukeCommand.bind("onNuke");
 		
-		crowdEnabledMessage = utilities.getMessage("crowdEnabled", "Mob %s enabled in world %s");
-		crowdDisabledMessage = utilities.getMessage("crowdDisable", "Mob %s disabled in world %s");
-		noWorldMessage = utilities.getMessage("noWorld", "Can't find world %s");
-		killedEntitiesMessage = utilities.getMessage("killedEntities", "Nuked %d of those darn %ss!");
-		noEntityMessage = utilities.getMessage("noEntities", "You are currently %s-free. Congrats!");
-		killFailedMessage = utilities.getMessage("killFailed", "Sorry, couldn't kill any &ss!");
+	    notControllingMessage = utilities.getMessage("notControlling", d.notControllingMessage);
+		crowdReplacedMessage = utilities.getMessage("crowdEnabled", d.crowdReplaceMessage);
+		crowdControlMessage = utilities.getMessage("crowdDisable", d.crowdControlMessage);
+		crowdReleasedMessage = utilities.getMessage("crowdDisable", d.crowdReleasedMessage);
+		noWorldMessage = utilities.getMessage("noWorld", d.noWorldMessage);
+		killedEntitiesMessage = utilities.getMessage("killedEntities", d.killedEntitiesMessage);
+		noEntityMessage = utilities.getMessage("noEntities", d.noEntityMessage);
+		killFailedMessage = utilities.getMessage("killFailed", d.killFailedMessage);
 	}
 	
-	public boolean onEnableCrowd(Player player, String[] parameters)
+	public ControlledWorld getWorldData(World world)
 	{
-		player.sendMessage("Not implemented yet!");
+		if (world == null) return null;
+		WorldData worldData = utilities.getWorld(getServer(), world);
+		
+		if (worldData == null) return null;
+		
+		ControlledWorld controlled = persistence.get(world, ControlledWorld.class);
+		if (controlled == null)
+		{
+			controlled = new ControlledWorld(worldData);
+			persistence.put(controlled);
+		}
+		
+		return controlled;
+	}
+	
+	public CreatureType getCreatureType(String name)
+	{
+		for (CreatureType ct: CreatureType.values())
+		{
+			if (ct.getName().equalsIgnoreCase(name)) return ct;
+		}
+		
+		return CreatureType.fromName(name);
+	}
+	
+	public boolean onControlCrowd(Player player, String[] parameters)
+	{
+		if (parameters.length < 1)
+		{
+			return false;
+		}
+		
+		CreatureType mobType = getCreatureType(parameters[0]);
+		
+		if (mobType == null)
+		{
+			unknownEntityMessage.sendTo(player, parameters[0]);
+			return true;
+		}
+		
+		ControlledWorld world = getWorldData(player.getWorld());
+		if (world == null)
+		{
+			noWorldMessage.sendTo(player, player.getWorld().getName());
+			return true;
+		}
+		
+		float percent = 1;
+		CreatureType targetType = null;
+		int currentRank = 1;
+		
+		if (parameters.length > 1)
+		{
+			try
+			{
+				percent = Float.parseFloat(parameters[1]);
+				if (percent > 1)
+				{
+					percent = (float)Integer.parseInt(parameters[1]) / 100;
+				}
+			} 
+			catch(NumberFormatException ex)
+			{
+			}
+		}
+		
+		if (parameters.length > 2)
+		{
+			targetType = getCreatureType(parameters[2]);
+		}
+		
+		List<ControlRule> rules = world.getRules();
+		if (rules == null)
+		{
+			rules = new ArrayList<ControlRule>();
+		}
+		
+		for (ControlRule rule : rules)
+		{
+			currentRank = rule.getRank() + 1;
+		}
+	
+		ControlRule newRule = new ControlRule(currentRank, mobType);
+		newRule.setPercentChance(percent);
+		newRule.setReplaceWith(targetType);
+		rules.add(newRule);
+		world.setRules(rules);
+		persistence.put(world);
+		if (percent == 0 || targetType == null)
+		{
+			crowdControlMessage.sendTo(player, mobType.getName());
+		}
+		else
+		{
+			crowdReplacedMessage.sendTo(player, mobType.getName(), targetType.getName(), (int)(percent * 100));
+		}
+		
 		return true;
 	}
 
-	public boolean onDisableCrowd(Player player, String[] parameters)
+	public boolean onReleaseCrowd(Player player, String[] parameters)
 	{
-		player.sendMessage("Not implemented yet!");
+		if (parameters.length < 1)
+		{
+			return false;
+		}
+		
+		CreatureType mobType = getCreatureType(parameters[0]);
+		
+		if (mobType == null)
+		{
+			unknownEntityMessage.sendTo(player, parameters[0]);
+			return true;
+		}
+		
+		ControlledWorld world = getWorldData(player.getWorld());
+		if (world == null)
+		{
+			noWorldMessage.sendTo(player, player.getWorld().getName());
+			return true;
+		}
+		
+		boolean modified = false;
+		List<ControlRule> rules = world.getRules();
+		List<ControlRule> newRules = new ArrayList<ControlRule>();
+		if (rules != null)
+		{
+			for (ControlRule rule : rules)
+			{
+				if (rule != null && rule.getCreatureType() != mobType)
+				{
+					newRules.add(rule);
+				}
+				else
+				{
+					modified = true;
+				}
+			}
+		}
+		
+		if (modified)
+		{
+			world.setRules(newRules);
+			persistence.put(world);
+			crowdReleasedMessage.sendTo(player, mobType.getName());
+		}
+		else
+		{
+			notControllingMessage.sendTo(player, mobType.getName());
+		}
+		
 		return true;
 	}
 	
@@ -104,14 +258,18 @@ public class CrowdControlPlugin extends JavaPlugin
 	{
 		String worldName = null;
 		WorldData targetWorld = null;
-		EntityType targetType = EntityType.GHAST;
+		CreatureType targetType = CreatureType.GHAST;
+		boolean nukeAll = false;
 		
 		if (parameters.length > 0)
 		{
-			targetType = EntityType.parseString(parameters[0]);
-			if (targetType == EntityType.UNKNOWN)
+			if (parameters[0].equalsIgnoreCase("all"))
 			{
-				targetType = EntityType.GHAST;
+				nukeAll = true;
+			}
+			else
+			{
+				targetType = CreatureType.fromName(parameters[0]);
 			}
 		}
 		
@@ -145,8 +303,7 @@ public class CrowdControlPlugin extends JavaPlugin
 			return true;
 		}
 		
-		Nuke nuke = new Nuke();
-		int killCount = nuke.nuke(world, targetType);
+		int killCount = controller.nuke(world, targetType, nukeAll);
 		
 		if (killCount > 0)
 		{
@@ -161,17 +318,22 @@ public class CrowdControlPlugin extends JavaPlugin
 	}
 	
 	protected final CrowdEntityListener listener = new CrowdEntityListener();
+	protected final Controller controller = new Controller();
 	
 	protected PluginCommand crowdCommand;
 	protected PluginCommand crowdControlCommand;
+	protected PluginCommand crowdReleaseCommand;
 	protected PluginCommand nukeCommand;
 
 	protected Message killedEntitiesMessage;
 	protected Message killFailedMessage;
 	protected Message noEntityMessage;
-	protected Message crowdEnabledMessage;
-	protected Message crowdDisabledMessage;
+	protected Message unknownEntityMessage;
+	protected Message crowdReplacedMessage;
+	protected Message crowdControlMessage;
+	protected Message crowdReleasedMessage;
 	protected Message noWorldMessage;
+	protected Message notControllingMessage;
 	
 	protected Persistence persistence = null;
 	protected PluginUtilities utilities = null;