+++ CrowdControl/src/main/java/com/elmakers/mine/bukkit/plugins/crowd/CrowdEntityListener.java
@@ -1,24 +1,28 @@
 package com.elmakers.mine.bukkit.plugins.crowd;
 
-import org.bukkit.entity.Entity;
-import org.bukkit.entity.Ghast;
 import org.bukkit.event.entity.CreatureSpawnEvent;
 import org.bukkit.event.entity.EntityListener;
 
+import com.elmakers.mine.bukkit.plugins.crowd.dao.ControlledWorld;
+import com.elmakers.mine.bukkit.plugins.persistence.Persistence;
+
 public class CrowdEntityListener extends EntityListener
 {
+	public void initialize(Persistence persistence, Controller controller)
+	{
+		this.persistence = persistence;
+		this.controller = controller;
+	}
+	
 	@Override
 	public void onCreatureSpawn(CreatureSpawnEvent event)
 	{
-		Entity entity = event.getEntity();
-		if (entity instanceof Ghast)
-		{
-			Ghast ghast = (Ghast)entity;
-			
-			// DIE MOFO, DIE!!!
-			ghast.setHealth(0);
-			
-			event.setCancelled(true);
-		}
+		ControlledWorld worldData = persistence.get(event.getLocation().getWorld().getName(), ControlledWorld.class);
+		if (worldData == null) return;
+		
+		controller.controlSpawnEvent(worldData, event);
 	}
+	
+	protected Persistence persistence;
+	protected Controller controller;
 }